requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-dotenv
torch
transformers
sentence-transformers
huggingface_hub
accelerate
pdfplumber
